Implemented better OOP
